requirements.txt
aiohttp
asyncpg
python-dotenv
aiosqlite
ccxt
fastapi
httpx
nats-py
numpy
pandas
pandas-ta
plotly
polars
prometheus-client
pydantic-settings
PyYAML
requests
scipy
streamlit
uvicorn
websockets
pytest
pytest-asyncio
pytest-mock
httpx
PyYAML
scipy

# Dev Tools
ruff
mypy
pre-commit
types-requests
types-PyYAML
tenacity
